auto refresh cookie

+++ utils/scheduler.py
@@ -1,11 +1,17 @@
 from datetime import datetime, timedelta
 
+from apscheduler.executors.asyncio import AsyncIOExecutor
 from apscheduler.executors.pool import ThreadPoolExecutor, ProcessPoolExecutor
 from apscheduler.jobstores.memory import MemoryJobStore
 import os
 
+from apscheduler.schedulers.asyncio import AsyncIOScheduler
 from apscheduler.schedulers.background import BackgroundScheduler
+from fastapi import Depends
 
+from dependences import get_cookie_service, get_proxy_service
+from service.cookie_abstract_service import CookieAbstractService
+from service.proxy_abstract_service import ProxyAbstractService
 from settings import POSIX_MEDIA_DIR, COOKIES_DIR
 
 jobstores = {
@@ -13,8 +19,7 @@ jobstores = {
 }
 
 executors = {
-    "default": ThreadPoolExecutor(max_workers=3),
-    "processpool": ProcessPoolExecutor(max_workers=1)
+    "default": AsyncIOExecutor()
 }
 
 job_defaults = {
@@ -22,18 +27,29 @@ job_defaults = {
     "max_instances": 1
 }
 
-scheduler = BackgroundScheduler(jobstores=jobstores, executors=executors, job_defaults=job_defaults)
+scheduler = AsyncIOScheduler(jobstores=jobstores, executors=executors, job_defaults=job_defaults)
 
-def clean_dir(path: str):
+async def clean_dir(path: str):
     files = os.listdir(path)
     for file in files:
+        if file == "cookie.txt":
+            continue
         file_datetime_created = datetime.strptime(file.split(".")[0], "%Y%m%d-%H%M%S")
         if datetime.now() - file_datetime_created > timedelta(hours=1):
             print(f"Deleting {file}")
             os.remove(os.path.join(path, file))
 
-def cleanup_task() -> None:
-    clean_dir(COOKIES_DIR)
-    clean_dir(POSIX_MEDIA_DIR)
+
+async def cleanup_task() -> None:
+    await clean_dir(COOKIES_DIR)
+    await clean_dir(POSIX_MEDIA_DIR)
+
+async def refresh_cookie_task(cookie_service: CookieAbstractService = get_cookie_service(), proxy_service : ProxyAbstractService = get_proxy_service()) -> None:
+    cookie_path = os.path.join(COOKIES_DIR, "cookie.txt")
+    proxy = await proxy_service.get_proxy()
+    print(proxy)
+    cookie_path = await cookie_service.refresh_cookie(proxy_url=proxy, cookie_path=cookie_path)
+    print(f"Refreshed cookie: {cookie_path}")
 
 scheduler.add_job(cleanup_task, "interval", minutes=60)
+scheduler.add_job(refresh_cookie_task, "interval", minutes=5)
